+++ src/InfoManager/KettleInfo.h
@@ -8,35 +8,35 @@
 typedef struct{
 	int heatState;
 	int pumpState;
-	int waterTemperature;
+	float waterTemperature;
 	int lidState;
 	int waterLevel;
 	int lockState;
 	int remainingTime;
-	int targetTemperature;
+	float targetTemperature;
 }info_t;
 
 
 /* setter */
 void setHeatState(int);
 void setPumpState(int);
-void setWaterTemperature(int);
+void setWaterTemperature(float);
 void setLidState(int);
 void setWaterLevel(int);
 void setLockState(int);
 void setRemainingTime(int);
-void setTargetTemperature(int);
+void setTargetTemperature(float);
 
 
 /* getter */
 int getHeatState(void);
 int getPumpState(void);
-int getWaterTemperature(void);
+float getWaterTemperature(void);
 int getLidState(void);
 int getWaterLevel(void);
 int getLockState(void);
 int getRemainingTime(void);
-int getTargetTemperature(void);
+float getTargetTemperature(void);
 
 
 //#endif
